second commit

+++ src/Employee.java
@@ -5,13 +5,13 @@ public class Employee {
 
     private int department;
 
-    private double salary;
+    private int salary;
 
     private static int count;
 
     private final int id;
 
-    public Employee(String fio,int department,double salary){
+    public Employee(String fio,int department,int salary){
             id=count++ +1;
             this.fio = fio;
             this.salary = salary;
@@ -31,7 +31,7 @@ public class Employee {
         return this.department;
     }
 
-    public double getSalary(){
+    public int getSalary(){
         return this.salary;
     }
 
@@ -39,7 +39,7 @@ public class Employee {
         this.department = department;
     }
 
-    public void setSalary(double salary){
+    public void setSalary(int salary){
         this.salary = salary;
     }
